Revert json viewer update
Since @uiw/react-json-view version 2.0.0-alpha.32 there have been changes to how the json data expansion works and in our case this is causing the data to always be collapsed, instead of using the configured collapsed level as previous. I tried changing the configuration but with no luck, therefore I will simply cap the version for now to the one before the changes.